requirements.txt
google-adk
google-genai
numpy
pandas
